Eliminated compilation warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21132 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/shader/shadercompiler/xmlshader/xmlshader.cpp
@@ -51,7 +51,7 @@ int csXMLShaderTech::textureUnits[shaderPass::TEXTUREMAX];
 size_t csXMLShaderTech::lastTexturesCount;
 
 csXMLShaderTech::csXMLShaderTech (csXMLShader* parent) : 
-  passes(0), passesCount(0), currentPass(~0),
+  passes(0), passesCount(0), currentPass((size_t)~0),
   xmltokens (parent->compiler->xmltokens)
 {
   csXMLShaderTech::parent = parent;
@@ -633,7 +633,7 @@ bool csXMLShaderTech::DeactivatePass ()
   if(currentPass>=passesCount)
     return false;
   shaderPass *thispass = &passes[currentPass];
-  currentPass = ~0;
+  currentPass = (size_t)~0;
 
   if(thispass->vproc) thispass->vproc->Deactivate ();
   if(thispass->vp) thispass->vp->Deactivate ();
